Fix table sorting bug
This provides a fix for the issue where some columns of the table were
not being sorted correctly. The issue is that some of the columns have
objects, so they would never be sorted differently when keyed into.
This adds a utility method for getting the internal value of an element
in the table, which lets us appropriately sort things like the
courseId.
Additionally changes to displaying N/A for missing dates in the table.

CR-2058