WIP: Add 'Create new enrollment' button
This adds in most of the frontend code needed to support the 'Create
new enrollment' button/action. An important thing to note is that the
API that we plan to use: /support/enrollment/{user} does not currently
support creating new enrollments, only updating them. To get this work
across the line, we'll have to update that API in the LMS.

CR-2068