feat: allow only whole dollar amounts in inputs for money
In the Support Tools: New Learner Credit Plan, a validity check was added that
the "Total account value" and the "Per learner spend cap" can only be numbers,
and no decimal point. If the user types any character other than a number,
the account value or the spend cap will not update on the screen.

ENT-6940